+++ src/config/index.ts
@@ -66,6 +66,7 @@ function getConfig(configDir: string, inputConfig: BDDInputConfig): BDDConfig {
     // important to resolve outputDir as it is used as unique key for input configs
     outputDir: resolveOutputDir(configDir, config.outputDir),
     importTestFrom: resolveImportTestFrom(configDir, config.importTestFrom),
+    watch: resolveWatch(configDir, config.watch),
   };
 }
 
@@ -101,6 +102,14 @@ function resolveImportTestFrom(
       varName,
     };
   }
+}
+
+function resolveWatch(configDir: string, watch: BDDInputConfig['watch']) {
+  if (!watch) return;
+  return {
+    extraPaths: watch.extraPaths?.map((watchPath) => path.resolve(configDir, watchPath)),
+    ignorePaths: watch.ignorePaths?.map((watchPath) => path.resolve(configDir, watchPath)),
+  };
 }
 
 function validateFeaturesRoot(featuresRoot?: string) {